Fix bug in DNSSD IPv6 address string conversion.
The old implementation converted each byte to a hexadecimal number
separately, meaning that leading zeros in a byte would be dropped
even in the middle of a group.

For example, the address 2620:149:110b:470e:0:0:0:e1a was incorrectly
rendered as              2620:149:11b:47e:0:0:0:e1a, with zeros dropped
in the 3rd and 4th octets.

I figured the most robust fix would be to use inet_ntop to format the
address, which produces a correctly abbreviated address like
2620:149:110b:470e::e1a. This is the approach used in DNSResolver_c-ares.swift.