Add public inits to record types
Motivation:

The record types are public but don't have public inits, this makes it
impossible for users to implement the `DNSResolver` protocol in a useful
way.

The record types are also `Equatable` but not `Hashable`. There's no
reason for them not to be `Hashable`.

Modifications:

- Add `public` `init`s to the record types
- Make the record types `Hashable`

Result:

Possible to implement a `DNSResolver`